move from_int_or_hex to ethers-core utils
